Added unit tests for Role methods and refactored unit tests and seperated them into individual classes for better navigation/readability
