+++ tma/src/main/java/com/sparta/tma/DAOs/RoleDAO.java
@@ -5,6 +5,9 @@ import com.sparta.tma.entities.Role;
 
 public class RoleDAO {
     public Role getRole(EmployeeDTO employeeDetails) {
+
+        if(employeeDetails.getRole() == null) throw new NullPointerException("Role not found: field is empty");
+
         String role = employeeDetails.getRole()
                 .trim()
                 .toUpperCase();
@@ -13,7 +16,7 @@ public class RoleDAO {
             case "ADMIN" -> Role.ADMIN;
             case "MANAGER" -> Role.MANAGER;
             case "EMPLOYEE" -> Role.EMPLOYEE;
-            default -> throw new IllegalArgumentException("Role not found");
+            default -> throw new IllegalArgumentException("Role not found, please enter a valid role (Admin/Manager/Employee)");
         };
     }
 }
